Added Thirst Was Taken's compatibility
With Thirst Was Taken installed, low thirst can now lower sanity, having a harsher effect on it than just being hungry. I also finally remembered to completely change the README to what it needs to be. Bug with the arrow overlay still needs to be fixed, but croissantnova and I are working to fix that.